Reworked the 'include forfeits'-option
Results are now structs as well and have a field showing if the result was a forfeit.

Forfeits will now be added and marked by a red background in the match table. By toggling the 'include forfeits'-option these results can be included/disregarded when calculating the new ratings (and stats like wins/losses.