Reworked the recalculateAllPeriods-method
RecalculateAllPeriods now (re-)creates Glicko-Players in the period their first result is in. The decay is now correctly applied.

Also fixed a bug in WinRatPer's updatePlayerID-method that led to the method not checking for the right displayAlias as well after finding the item with the ID to update.

Added a 'description' column in the 'Match Report'-tab that will show the tourney name, when imported via challonge.com. Added a field for a description in the Result-struct to save and load those.

Resized tables