Put the 'include forfeits in rating'-option to the top.
Now the options are descending in probability to get changed. With API key having to be applied, 'include forfeits' being subjective and the starting values probably best left unchanged. Tau might be debatably more probable to be changed, but its influence is hardly noticeable in the given limits (plus, it only affects one of the rating values over time).