Added statistics to 'Player Edit'-tab. Revised the updatePlayerEntry-method.
Showing the statistics in the 'Player Edit'-tab makes it easier for the user to review what player to show and hide in the ranking table. It is also the only way for the user to see the deviation and volatility values without looking into the save-files.

The updatePlayerEntry-method was a little specific, so I decided to change it to getPlayerStats, which retrieves the stats used to update the player-entries in the rating table. This method is way more useful being used for the 'Player Edit'-tab's statistics as well.