Removing results does not require the winning player to be saved as player 1 in results.json now.
Dennis-clause added. When a result is entered in challonge as being tied by score (1-1 for example), but a winner is chosen, the result is added as being won by the chosen winner by forfeit. This is because there was a case where someone reported a set like that, after his opponent had to leave, when the score was 1-1.

Loading the world at startup will only read the starting rating values of each player now, but add enough entries for every rating period.